deleting formatic wallet and fixing tests (#517)
* deleting formattic wallet and fixing tests

* fixing formatic

* enum fix and activating tests
