+++ src/components/modals/WalletModal/index.tsx
@@ -6,8 +6,7 @@ import { URI_AVAILABLE } from '@anxolin/walletconnect-connector'
 import { UnsupportedChainIdError, useWeb3React } from '@web3-react/core'
 import ReactGA from 'react-ga'
 
-import { fortmatic, injected, walletconnect } from '../../../connectors'
-import { OVERLAY_READY } from '../../../connectors/Fortmatic'
+import { injected, walletconnect } from '../../../connectors'
 import { SUPPORTED_WALLETS } from '../../../constants'
 import usePrevious from '../../../hooks/usePrevious'
 import { useWalletModalOpen, useWalletModalToggle } from '../../../state/application/hooks'
@@ -179,12 +178,6 @@ const WalletModal: React.FC = () => {
       }
     }
   }
-
-  useEffect(() => {
-    fortmatic.on(OVERLAY_READY, () => {
-      toggleWalletModal()
-    })
-  }, [toggleWalletModal])
 
   const getOptions = () => {
     const isMetamask = window.ethereum && window.ethereum.isMetaMask
