+++ src/connectors/index.ts
@@ -6,13 +6,11 @@ import { WalletLinkConnector } from '@web3-react/walletlink-connector'
 
 import {
   CHAIN_ID,
-  FORTMATIC_KEY,
   NETWORK_URL_MAINNET,
   NETWORK_URL_RINKEBY,
   NETWORK_URL_XDAI,
   PORTIS_ID,
 } from '../constants/config'
-import { FortmaticConnector } from './Fortmatic'
 
 interface NetworkConnectorArguments {
   urls: { [chainId: number]: string }
@@ -65,12 +63,6 @@ export const walletconnect = {
     pollingInterval: POLLING_INTERVAL,
   }),
 }
-
-// mainnet only
-export const fortmatic = new FortmaticConnector({
-  apiKey: FORTMATIC_KEY,
-  chainId: 1,
-})
 
 // mainnet only
 export const portis = new PortisConnector({
