+++ src/state/orderPlacement/hooks.ts
@@ -8,7 +8,6 @@ import { useDispatch, useSelector } from 'react-redux'
 
 import { additionalServiceApi } from '../../api'
 import { ClearingPriceAndVolumeData } from '../../api/AdditionalServicesApi'
-import { EASY_AUCTION_NETWORKS } from '../../constants'
 import easyAuctionABI from '../../constants/abis/easyAuction/easyAuction.json'
 import { NUMBER_OF_DIGITS_FOR_INVERSION } from '../../constants/config'
 import { useActiveWeb3React } from '../../hooks'
@@ -18,7 +17,7 @@ import { AuctionInfoDetail, useAuctionDetails } from '../../hooks/useAuctionDeta
 import { useGetClaimInfo } from '../../hooks/useClaimOrderCallback'
 import { useContract } from '../../hooks/useContract'
 import { useClearingPriceInfo } from '../../hooks/useCurrentClearingOrderAndVolumeCallback'
-import { ChainId, getTokenDisplay } from '../../utils'
+import { ChainId, EASY_AUCTION_NETWORKS, getTokenDisplay } from '../../utils'
 import { getLogger } from '../../utils/logger'
 import { convertPriceIntoBuyAndSellAmount, getInverse } from '../../utils/prices'
 import { AppDispatch, AppState } from '../index'
