+++ src/utils/index.test.ts
@@ -1,13 +1,11 @@
-import { JSBI, Percent, Token, TokenAmount } from 'uniswap-xdai-sdk'
+import { JSBI, Percent } from 'uniswap-xdai-sdk'
 
 import { BigNumber } from '@ethersproject/bignumber'
-import { AddressZero } from '@ethersproject/constants'
 
 import {
   ChainId,
   basisPointsToPercent,
   calculateGasMargin,
-  calculateSlippageAmount,
   getExplorerLink,
   isAddress,
   shortenAddress,
@@ -28,30 +26,6 @@ describe('utils', () => {
       expect(getExplorerLink(ChainId.RINKEBY, 'abc', 'address')).toEqual(
         'https://rinkeby.etherscan.io/address/abc',
       )
-    })
-  })
-
-  describe('#calculateSlippageAmount', () => {
-    it('bounds are correct', () => {
-      const tokenAmount = new TokenAmount(new Token(ChainId.MAINNET, AddressZero, 0), '100')
-      expect(() => calculateSlippageAmount(tokenAmount, -1)).toThrow()
-      expect(calculateSlippageAmount(tokenAmount, 0).map((bound) => bound.toString())).toEqual([
-        '100',
-        '100',
-      ])
-      expect(calculateSlippageAmount(tokenAmount, 100).map((bound) => bound.toString())).toEqual([
-        '99',
-        '101',
-      ])
-      expect(calculateSlippageAmount(tokenAmount, 200).map((bound) => bound.toString())).toEqual([
-        '98',
-        '102',
-      ])
-      expect(calculateSlippageAmount(tokenAmount, 10000).map((bound) => bound.toString())).toEqual([
-        '0',
-        '200',
-      ])
-      expect(() => calculateSlippageAmount(tokenAmount, 10001)).toThrow()
     })
   })
